handle null returned from listFiles when trying to clean mirror
Merge-request: TC-MR-15257
Merged-by: Ilya Voronin <ilya.voronin@jetbrains.com>

+++ git-agent/src/main/java/jetbrains/buildServer/buildTriggers/vcs/git/agent/FSImpl.java
@@ -26,8 +26,13 @@ public class FSImpl implements FS {
       return false;
     }
 
+
+    File[] files = dir.listFiles();
+    if (files == null) {
+      return false;
+    }
     boolean successfull = true;
-    for (File file : Objects.requireNonNull(dir.listFiles())) {
+    for (File file : files) {
       successfull &= FileUtil.delete(file);
     }
     return successfull;
